CHROMIUM: iwl7000: pcie: clarify the used/empty/free list counters
Currently the used_count counts the number of buffers that a q needs the
allocator to allocate for it. These buffers could sit in different lists
- sometimes in the rx_used list of the q, and sometimes in the rx_empty
list of the allocator. This could cause some confusion when trying to
understand what's going on in this flow.

Clear it up a little by changing the meaning of used_count to simply
reflect the number of buffers in the rx_used list at any given moment,
and add a new counter for keeping track of how many times 8-buffer lists
were given to the allocator by this q, so it could keep track of how
many times it could attempt to get allocated buffers back from it.

Change-Id: Ibea8f6f8707d27207994a69d589f5061c1cfbf58
Signed-off-by: Naftali Goldstein <naftali.goldstein@intel.com>
iwl7000-tree: ee14a915f90e79d5757942baa76a2e85b3458f20