CHROMIUM: iwl7000: mvm: synchronize TID queue removal
When we mark a TID as no longer having a queue, there's no
guarantee the TX path isn't using this txq_id right now,
having accessed it just before we reset the value. To fix
this, add synchronize_net() when we change the TIDs from
having a queue to not having one, so that we can then be
sure that the TX path is no longer accessing that queue.

Change-Id: I5ea0cc51e22f0d94b6c173a7458abac9bbcb895f
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 0f49626b09576e933117b075dab6f3355e18e017