CHROMIUM: iwl7000: mac80211: Fix station bandwidth setting after channel switch
When performing a channel switch flow for a managed interface, the
flow did not update the bandwidth of the AP station and the rate
scale algorithm. In case of a channel width downgrade, this would
result with the rate scale algorithm using a bandwidth that does not
match the interface channel configuration.

Fix this by updating the AP station bandwidth and rate scaling algorithm
before the actual channel change in case of a bandwidth downgrade, or
after the actual channel change in case of a bandwidth upgrade.

Change-Id: I9e1049f955d9bd2623d50a12b72b74d480ac4df5
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: ddfc2522d3bfc811bef89018a16a1fa101e0a3f0