CHROMIUM: iwl7000: mvm: FTM: make the macaddr_random flag debugfs setable
If this flag is set, the FW will randomize the last byte, even if the
host supplied a mask that has no randomization.
In order to get fully static mac address, this flag should be unset.

Currently this flag is always set before sending the command. Change this
by getting the value that is stored in the driver the same way that it
is done for other fields. This stored value can be changed using debugfs.
Keep the default behaviour to be set.

Change-Id: Ib97d4523f4734781c4e9450cdce6cbc9b817ca9f
Signed-off-by: Beni Lev <beni.lev@intel.com>
iwl7000-tree: 7af437a746bb9e10142ec3e77b7c7e4953b44644