CHROMIUM: iwl7000: mvm: fix setting HE ppe FW config
The FW expects to get the ppe value for each NSS-BW pair in the same
format as in the he phy capabilities IE, which means that a value of 0
implies ppe should be used for BPSK (mcs 0). If there are no PPE
thresholds in the IE, or if for some NSS-RU pair there's no threshold
set for it (this could happen because it's a variable-sized field), it
means no PPE should not be used for that pair, so the value sent to FW
should be 7 which corresponds to "none".

Change-Id: I86aa519afbce41ceb2e4da6f50d2ddcfe579b353
Signed-off-by: Naftali Goldstein <naftali.goldstein@intel.com>
iwl7000-tree: 3ba0ebc557c67d07bae51b60e5f2cc8c7a72f292