CHROMIUM: iwl7000: wrt: add rt status and num of rx/tx fifos to dump
Add the rt status of the last assert or 0 if the dump collection
was not initiated by an assert.
Add number of rx and tx fifos in use.
the fields are added to dump info lst file.

Change-Id: I9b47afb97ee97f53795c25e6eafccd3783ad2b85
Signed-off-by: Shahar S Matityahu <shahar.s.matityahu@intel.com>
iwl7000-tree: 6d774007e67c65fde4526b28b5d5af03f5e0e714