CHROMIUM: mwifiex: advertise NL80211_FEATURE_LOW_PRIORITY_SCAN
It turns out this version of the driver *does* support
LOW_PRIORITY_SCAN, but it's just hidden by the fact that it uses a
downstream version called CFG80211_SCAN_FLAG_TX_ABORT /
NL80211_SCAN_FLAG_TX_ABORT, with the same values. But it's missing the
feature flag, so it got missed in CL:1185523 ("Revert "CHROMIUMOS:
Ignore driver's feature flag for low priority scan"").

Introduce the feature flag as it appears upstream, so wpa_supplicant
knows how to look for it again.

BUG=chromium:893857
TEST=`network_WiFi_BgscanBackoff` tests

Change-Id: I697838eb02646029481e7487ce2a3b86e6a62361
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1298371
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>