CHROMIUM: iwl7000: mac80211: never pass NULL params to ieee80211_if_add()
This isn't really a problem now, but it means that the function
has a few NULL checks that are only relevant when coming from
the initial interface added in mac80211, and that's confusing.
Just pass non-NULL (but equivalently empty) in that case and
remove all the NULL checks.

Change-Id: I3942f571879d300cc8513ddbdf59dc16b919dd51
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: b819ddfa0a7f66d175bd1050723a6474c41c1d49