CHROMIUM: iwl7000: mvm: freeze management TXQ timer when station goes to sleep
We maintain a timer for each Tx queue to detect stalls and
be able to recover / debug.
When we work in AP mode, we can freeze the Tx queue timer if
a station goes to sleep, because we don't want to warn about
stalls that are caused by faulty clients that don't wake up
on time.
This mechanism was applied to the queues of the clients, but
the management queue was omitted. Fix that.

Change-Id: I51a1ca81b3c0463fd9e8c3f96173e31ce28f8e0d
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: f3ef018aced58d6f99f031aa98eae8c8920d0faf