CHROMIUM: iwl7000: mvm: Do not return an error value on HW restart
When HW restart is requested but not started yet, commands would not
be sent to the FW, and some function calls would return an error. In
case of iwl_mvm_mac_sta_state() returning an error value when a
station is removed can lead to an unneeded warning in
__sta_info_destroy_part2().

Handle this by setting the return value to 0, in case HW restart is
in progress.

Change-Id: Ib830d4b4e726351f8494ac7988e5377c49d7f770
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 37289b39fd182bc8ca740eb87a02857a5fa5f1b2